requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121    # adapt to your CUDA version

flask
langchain==0.3.23
langchain-community==0.3.16
sentence-transformers
faiss-cpu
docx2txt
pdfminer.six
torch==2.5.1
vllm==0.6.4
gunicorn